language_bot/requirements.txt
# LanguageBot requirements
discord.py>=2.4.0
python-dotenv>=1.0.1
httpx>=0.27.0
langdetect>=1.0.9
openai>=1.12.0
